it's working -- using variadic tuple from agency, various fixes related to tuple_of_iterator_references, variadic or_ and_

+++ thrust/detail/tuple/tuple.h
@@ -53,7 +53,7 @@ template <
 class tuple;
 
 // forward declaration of tuple_element
-template<int i, typename T> struct tuple_element;
+template<size_t i, typename T> struct tuple_element;
 
 // specializations for tuple_element
 template<class T>
